Use size_t more consistently

+++ src/kalypsso/core/eos/MieGruneisenEosArray.cpp
@@ -186,7 +186,7 @@ MieGruneisenEosArray<device_t>::MieGruneisenEosArray(ConfigMap const & config_ma
 // =====================================================================
 template <typename device_t>
 KOKKOS_FUNCTION real_t
-MieGruneisenEosArray<device_t>::material_gruneisen_param(int i_mat, real_t rho) const
+MieGruneisenEosArray<device_t>::material_gruneisen_param(size_t i_mat, real_t rho) const
 {
   // get EOS type
   const auto eos_type = MG_EOS_TYPE::_from_integral_unchecked(m_material_eos_type(i_mat));
@@ -231,7 +231,7 @@ MieGruneisenEosArray<device_t>::material_gruneisen_param(int i_mat, real_t rho) 
 // =====================================================================
 template <typename device_t>
 KOKKOS_FUNCTION real_t
-MieGruneisenEosArray<device_t>::material_specific_eint_ref(int i_mat, real_t rho) const
+MieGruneisenEosArray<device_t>::material_specific_eint_ref(size_t i_mat, real_t rho) const
 {
   // get EOS type
   const auto eos_type = MG_EOS_TYPE::_from_integral_unchecked(m_material_eos_type(i_mat));
@@ -276,7 +276,7 @@ MieGruneisenEosArray<device_t>::material_specific_eint_ref(int i_mat, real_t rho
 // =====================================================================
 template <typename device_t>
 KOKKOS_FUNCTION real_t
-MieGruneisenEosArray<device_t>::material_pressure_ref(int i_mat, real_t rho) const
+MieGruneisenEosArray<device_t>::material_pressure_ref(size_t i_mat, real_t rho) const
 {
   // get EOS type
   const auto eos_type = MG_EOS_TYPE::_from_integral_unchecked(m_material_eos_type(i_mat));
@@ -321,7 +321,7 @@ MieGruneisenEosArray<device_t>::material_pressure_ref(int i_mat, real_t rho) con
 // =====================================================================
 template <typename device_t>
 KOKKOS_FUNCTION real_t
-MieGruneisenEosArray<device_t>::material_sound_speed_square(int    i_mat,
+MieGruneisenEosArray<device_t>::material_sound_speed_square(size_t i_mat,
                                                             real_t pressure,
                                                             real_t rho) const
 {
@@ -368,7 +368,9 @@ MieGruneisenEosArray<device_t>::material_sound_speed_square(int    i_mat,
 // =====================================================================
 template <typename device_t>
 KOKKOS_FUNCTION real_t
-MieGruneisenEosArray<device_t>::material_bulk_modulus(int i_mat, real_t pressure, real_t rho) const
+MieGruneisenEosArray<device_t>::material_bulk_modulus(size_t i_mat,
+                                                      real_t pressure,
+                                                      real_t rho) const
 {
 
   // const auto c2 = material_sound_speed_square(i_mat, pressure, rho);
@@ -417,7 +419,7 @@ MieGruneisenEosArray<device_t>::material_bulk_modulus(int i_mat, real_t pressure
 // =====================================================================
 template <typename device_t>
 KOKKOS_FUNCTION real_t
-MieGruneisenEosArray<device_t>::material_specific_eint_from_pressure(int    i_mat,
+MieGruneisenEosArray<device_t>::material_specific_eint_from_pressure(size_t i_mat,
                                                                      real_t pressure,
                                                                      real_t rho) const
 {
@@ -465,7 +467,7 @@ MieGruneisenEosArray<device_t>::material_specific_eint_from_pressure(int    i_ma
 // =====================================================================
 template <typename device_t>
 KOKKOS_FUNCTION real_t
-MieGruneisenEosArray<device_t>::material_pressure_from_specific_eint(int    i_mat,
+MieGruneisenEosArray<device_t>::material_pressure_from_specific_eint(size_t i_mat,
                                                                      real_t eint_specific,
                                                                      real_t rho) const
 {
